Implements the functionality required for a real CSW server.
Adds a paster command which will then interact with a locally installed
pycsw (v1.8.0) installation.  The paster command can be used to load and
clear a pycsw specific postgres database (ideally postgis database) but
the actual server itself is provided by pycsw (via wsgi).  pycsw
interacts with ckan via the API and so doesn't have to be on the same
machine or in the same virtual environment.