[1895] Failing test for issue
Extend the gemini-waf test to look at the number of resources produced by the
harvest. Currently the final assertion of this test will fail because each time
the harvester is run it will create a new set of resources because it does not
check to see if the resource is already present.